The error with the restriction in colors is solved

+++ src/main/java/Engine/Parser/Parser.java
@@ -26,9 +26,9 @@ public class Parser {
             var x = Integer.parseInt(positionAndColor[0].split(",")[0]);
             var y = Integer.parseInt(positionAndColor[0].split(",")[1]);
             var stringColor = positionAndColor[1].split(",");
-            var r = Integer.parseInt(stringColor[0]);
-            var g = Integer.parseInt(stringColor[1]);
-            var b = Integer.parseInt(stringColor[2]);
+            var r = Double.parseDouble(stringColor[0]);
+            var g = Double.parseDouble(stringColor[1]);
+            var b = Double.parseDouble(stringColor[2]);
             var color = new Color(r, g, b, 1);
             var pixel = new Pixel(x, y, color);
             result.add(pixel);
